Move coverage and reporter to root Jest config
